requirements.txt
ctransformers==0.2.27
faiss-cpu==1.7.4
fastapi=0.104.0
langchain==0.0.323
pydantic==2.4.2
python-dotenv==1.0.0
sentence-transformers==2.2.2
uvicorn=0.23.2
unstructured==0.10.26